requirements.txt
asgiref==3.2.5
Django==3.0.4
django-bootstrap-form==3.4
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
